Expanded match() and findNext() to match the most common use cases.
match() and findNext() now are overloaded -- one for strings, one for
character arrays. Templates were not an options, sadly, because these
are virtual functions.

I am beginning to question the architecture in place -- do I really need
a fully functional DFA implementation? Might it be wiser to always
normalize during a build, so the only class that's public-facing is the
NFA? It might make certain things easier.